Prototype notebook update
The prototype notebook now contains an example of how to do a hydro projection based on climate data. It uses gcm data from the four available rcp scenarios and simulates hydro outputs for two glaciers in Asia. On/off glacier melt and liquid precipitation is combined to get the total runoff. One can already "see" the occurence of peak water.