Glaciated basin analysis
First try at simulating the glacier runoff in a whole bains. This works on the Rofental, simulates all its glaciers 2020-2100 and compiles the output to a file. Some utility functions that can later be put into a module for re-use. Also contains code to create a plot showing the runoff evolution.